Add estimate diff size test with fork switches
