+++ crates/evm/src/tests/queries/estimate_gas_tests.rs
@@ -8,10 +8,7 @@ use reth_primitives::BlockNumberOrTag;
 use reth_rpc_eth_types::RpcInvalidTransactionError;
 use serde_json::json;
 use sov_modules_api::default_context::DefaultContext;
-use sov_modules_api::fork::Fork;
-use sov_modules_api::hooks::HookSoftConfirmationInfo;
 use sov_modules_api::{Spec, WorkingSet};
-use sov_rollup_interface::spec::SpecId as SovSpecId;
 
 use crate::query::MIN_TRANSACTION_GAS;
 use crate::smart_contracts::{CallerContract, SimpleStorageContract};
@@ -308,153 +305,6 @@ fn test_tx_request_fields_gas_fork1() {
             gas_used: U256::from_str("0x6e67").unwrap()
         }
     );
-}
-
-#[test]
-fn test_access_list_with_forks() {
-    // 0x819c5497b157177315e1204f52e588b393771719 -- Storage contract
-    // 0x5ccda3e6d071a059f00d4f3f25a1adc244eb5c93 -- Caller contract
-
-    let (mut evm, mut working_set, signer, _) = init_evm_with_caller_contract();
-
-    let mut l2_height = 2;
-
-    let fork_fn = |num: u64| {
-        if num < 3 {
-            Fork::new(SovSpecId::Genesis, 0)
-        } else {
-            Fork::new(SovSpecId::Fork1, 4)
-        }
-    };
-
-    let caller = CallerContract::default();
-    let input_data = caller.call_set_call_data(
-        Address::from_str("0x819c5497b157177315e1204f52e588b393771719").unwrap(),
-        42,
-    );
-
-    let tx_req_contract_call = TransactionRequest {
-        from: Some(signer.address()),
-        to: Some(TxKind::Call(address!(
-            "5ccda3e6d071a059f00d4f3f25a1adc244eb5c93"
-        ))),
-        gas: Some(10000000),
-        gas_price: Some(100),
-        max_fee_per_gas: None,
-        max_priority_fee_per_gas: None,
-        value: None,
-        input: TransactionInput::new(input_data.into()),
-        nonce: Some(3u64),
-        chain_id: Some(1u64),
-        access_list: None,
-        max_fee_per_blob_gas: None,
-        blob_versioned_hashes: None,
-        transaction_type: None,
-        sidecar: None,
-        authorization_list: None,
-    };
-
-    let no_access_list = evm.eth_estimate_gas_inner(
-        tx_req_contract_call.clone(),
-        None,
-        &mut working_set,
-        fork_fn,
-    );
-    assert_eq!(no_access_list.unwrap(), U256::from_str("0x54ef").unwrap());
-
-    let form_access_list = evm.create_access_list_inner(
-        tx_req_contract_call.clone(),
-        None,
-        &mut working_set,
-        fork_fn,
-    );
-
-    assert_eq!(
-        form_access_list.unwrap(),
-        AccessListWithGasUsed {
-            access_list: AccessList(vec![]),
-            gas_used: U256::from_str("0x54ef").unwrap()
-        }
-    );
-
-    let tx_req_with_access_list = TransactionRequest {
-        access_list: Some(AccessList(vec![])),
-        ..tx_req_contract_call.clone()
-    };
-
-    let with_access_list =
-        evm.eth_estimate_gas_inner(tx_req_with_access_list, None, &mut working_set, fork_fn);
-    assert_eq!(with_access_list.unwrap(), U256::from_str("0x54ef").unwrap());
-
-    let soft_confirmation_info = HookSoftConfirmationInfo {
-        l2_height,
-        da_slot_hash: [1u8; 32],
-        da_slot_height: 1,
-        da_slot_txs_commitment: [42u8; 32],
-        pre_state_root: [10u8; 32].to_vec(),
-        current_spec: SovSpecId::Fork1,
-        pub_key: vec![],
-        deposit_data: vec![],
-        l1_fee_rate: 0,
-        timestamp: 0,
-    };
-
-    evm.begin_soft_confirmation_hook(&soft_confirmation_info, &mut working_set);
-    evm.end_soft_confirmation_hook(&soft_confirmation_info, &mut working_set);
-    evm.finalize_hook(&[99u8; 32].into(), &mut working_set.accessory_state());
-
-    let no_access_list = evm.eth_estimate_gas_inner(
-        tx_req_contract_call.clone(),
-        None,
-        &mut working_set,
-        fork_fn,
-    );
-    // After fork the gas cost is different.
-    assert_eq!(no_access_list.unwrap(), U256::from_str("0x788c").unwrap());
-
-    let form_access_list = evm.create_access_list_inner(
-        tx_req_contract_call.clone(),
-        None,
-        &mut working_set,
-        fork_fn,
-    );
-
-    assert_eq!(
-        form_access_list.unwrap(),
-        AccessListWithGasUsed {
-            access_list: AccessList(vec![AccessListItem {
-                address: address!("819c5497b157177315e1204f52e588b393771719"),
-                storage_keys: vec![b256!(
-                    "0000000000000000000000000000000000000000000000000000000000000000"
-                )]
-            }]),
-            gas_used: U256::from_str("0x775e").unwrap()
-        }
-    );
-
-    let tx_req_with_access_list = TransactionRequest {
-        access_list: Some(AccessList(vec![AccessListItem {
-            address: address!("819c5497b157177315e1204f52e588b393771719"),
-            storage_keys: vec![b256!(
-                "0000000000000000000000000000000000000000000000000000000000000000"
-            )],
-        }])),
-        ..tx_req_contract_call.clone()
-    };
-
-    let with_access_list =
-        evm.eth_estimate_gas_inner(tx_req_with_access_list, None, &mut working_set, fork_fn);
-    assert_eq!(with_access_list.unwrap(), U256::from_str("0x775e").unwrap());
-
-    // Get pre fork gas cost and expect it to still work same
-
-    let no_access_list = evm.eth_estimate_gas_inner(
-        tx_req_contract_call.clone(),
-        Some(BlockNumberOrTag::Number(2)),
-        &mut working_set,
-        fork_fn,
-    );
-    assert_eq!(no_access_list.unwrap(), U256::from_str("0x54ef").unwrap());
 }
 
 #[test]
